# lib/api-spec/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
  description: WorkforceOS - Multilingual Workforce Management Platform API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        employeeCode:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        salary:
          type: string
          nullable: true
        joinDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, on-leave]
        avatar:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [employeeCode, fullName, email]
      properties:
        employeeCode:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        salary:
          type: string
        joinDate:
          type: string
        departmentId:
          type: integer
        status:
          type: string
          enum: [active, inactive, on-leave]
        address:
          type: string

    # Department
...
    Leave:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        type:
          type: string
          enum: [annual, sick, emergency, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        daysCount:
          type: integer
        reason:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: integer
          nullable: true
        approvedAt:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    LeaveInput:
      type: object
      required: [employeeId, type, startDate, endDate, daysCount]
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [annual, sick, emergency, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        daysCount:
          type: integer
        reason:
          type: string
    LeaveUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: integer

    # Schedule
    Schedule:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        date:
          type: string
        shiftType:
          type: string
          enum: [morning, evening, night, off]
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        isRemote:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ScheduleInput:
      type: object
      required: [employeeId, date, shiftType]
      properties:
        employeeId:
          type: integer
        date:
          type: string
        shiftType:
          type: string
          enum: [morning, evening, night, off]
        startTime:
          type: string
        endTime:
          type: string
        isRemote:
          type: boolean
        notes:
          type: string

    # Payroll
    Payroll:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        period:
          type: string
        basicSalary:
          type: string
        overtime:
          type: string
        bonus:
          type: string
        deductions:
          type: string
        tax:
          type: string
        netSalary:
          type: string
        status:
          type: string
          enum: [pending, processing, paid]
        paidAt:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PayrollUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, paid]

    # WorkRequest
    WorkRequest:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        type:
          type: string
          enum: [leave, overtime, shift_change, expense, equipment]
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: integer
          nullable: true
        approvedAt:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    WorkRequestInput:
      type: object
      required: [employeeId, type, title]
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [leave, overtime, shift_change, expense, equipment]
        title:
          type: string
        description:
          type: string
    WorkRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: integer

    # Asset
...
    EmployeeList:
      type: object
      properties:
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        total:
          type: integer
...
    LeaveList:
      type: object
      properties:
        leaves:
          type: array
          items:
            $ref: '#/components/schemas/Leave'
        total:
          type: integer
    ScheduleList:
      type: object
      properties:
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
    PayrollList:
      type: object
      properties:
        payroll:
          type: array
          items:
            $ref: '#/components/schemas/Payroll'
        total:
          type: integer
    WorkRequestList:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/WorkRequest'
        total:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      operationId: getEmployees
      summary: List employees
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeList'
    post:
      operationId: createEmployee
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
  /employees/{id}:
    get:
      operationId: getEmployee
      summary: Get employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
    put:
      operationId: updateEmployee
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
    delete:
      operationId: deleteEmployee
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No content

  # Departments
...
  /leaves:
    get:
      operationId: getLeaves
      summary: List leaves
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveList'
    post:
      operationId: createLeave
      summary: Create leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leave'
  /leaves/{id}:
    put:
      operationId: updateLeave
      summary: Update leave status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leave'
    delete:
      operationId: deleteLeave
      summary: Delete leave
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No content

  # Schedules
  /schedules:
    get:
      operationId: getSchedules
      summary: List schedules
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleList'
    post:
      operationId: createSchedule
      summary: Create schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'

  # Payroll
  /payroll:
    get:
      operationId: getPayroll
      summary: List payroll records
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayrollList'
  /payroll/{id}:
    put:
      operationId: updatePayroll
      summary: Update payroll status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayrollUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payroll'

  # Work Requests
  /requests:
    get:
      operationId: getRequests
      summary: List work requests
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRequestList'
    post:
      operationId: createRequest
      summary: Create work request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkRequestInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRequest'
  /requests/{id}:
    put:
      operationId: updateRequest
      summary: Update request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkRequestUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRequest'

  # Assets
...