# lib/api-zod/src/generated/types/leaveList.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WorkforceOS - Multilingual Workforce Management Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Leave } from './leave';

export interface LeaveList {
  leaves?: Leave[];
  total?: number;
}