# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WorkforceOS - Multilingual Workforce Management Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { ScheduleInputShiftType } from './scheduleInputShiftType';

export interface ScheduleInput {
  employeeId: number;
  date: string;
  shiftType: ScheduleInputShiftType;
  startTime?: string;
  endTime?: string;
  isRemote?: boolean;
  notes?: string;
}